minor rearranging
