+++ src/swell/swell.py
@@ -10,9 +10,9 @@
 
 import click
 
-from swell.deployment.prep_config import clone_config, prepare_config
-from swell.deployment.prep_config_utils import get_platforms
-from swell.deployment.create_experiment import create_experiment_directory
+from swell.deployment.platforms.platforms import get_platforms
+from swell.deployment.create_experiment import clone_config, create_experiment_directory
+from swell.deployment.create_experiment import prepare_config
 from swell.deployment.launch_experiment import launch_experiment
 from swell.tasks.base.task_base import task_wrapper, get_tasks
 from swell.test.test_driver import test_wrapper, valid_tests
